[advanced] useFetchEvents 훅 분리
